refactor: askmark 네이밍 변경 + 스타일링 조정
