+++ src/components/ToolTip/index.jsx
@@ -1,8 +1,9 @@
-import React from "react";
-import styled from "styled-components";
+import React from 'react';
+import styled from 'styled-components';
 
 const Container = styled.div`
   display: flex;
+  position: relative;
   width: fit-content;
   gap: 10px;
   text-align: center;
@@ -18,7 +19,7 @@ const Container = styled.div`
   }
 `;
 
-const AskMarkContainer = styled.button`
+const ToolTipContainer = styled.button`
   width: 27px;
   height: 27px;
   border: 1px solid #bababa;
@@ -39,6 +40,8 @@ const AskMarkContainer = styled.button`
 
 const InformationContainer = styled.div`
   display: none;
+  position: absolute;
+  left: 40px;
   width: 180px;
   height: fit-content;
   font-size: 12px;
@@ -46,10 +49,10 @@ const InformationContainer = styled.div`
   word-break: keep-all;
 `;
 
-const AskMark = () => {
+const ToolTip = () => {
   return (
     <Container>
-      <AskMarkContainer>?</AskMarkContainer>
+      <ToolTipContainer>?</ToolTipContainer>
       <InformationContainer>
         카드 뒷면 서명란에 인쇄되어 있는 19자리 중 마지막 3자리 숫자 입니다
       </InformationContainer>
@@ -57,4 +60,4 @@ const AskMark = () => {
   );
 };
 
-export default AskMark;
+export default ToolTip;
