+++ 피드백반영내용.md
@@ -4,9 +4,9 @@
 
 - [x] styled component 상단으로 위치를 변경하기
 
-- [ ] AskMark, Tooltip으로 네이밍 변경하기
+- [x] AskMark, Tooltip으로 네이밍 변경하기
 
-- [ ] AskMark의 InformationContainer의 포지션 absolute로 바꾸기
+- [x] AskMark의 InformationContainer의 포지션 absolute로 바꾸기
 
 - [ ] Button 컴포넌트 네이밍 변경하기
